+++ program_youtube_downloader/__init__.py
@@ -1,4 +1,10 @@
 """Public exports for program_youtube_downloader."""
 from .exceptions import DownloadError, ValidationError
+from .utils import clear_screen, program_break_time
 
-__all__ = ["DownloadError", "ValidationError"]
+__all__ = [
+    "DownloadError",
+    "ValidationError",
+    "clear_screen",
+    "program_break_time",
+]
