+++ program_youtube_downloader/cli_utils.py
@@ -11,7 +11,7 @@ from .constants import (
     TITLE_QUESTION_MENU_ACCUEIL,
     CHOICE_MENU_ACCUEIL,
 )
-from .youtube_downloader import program_break_time, clear_screen
+from .utils import program_break_time, clear_screen
 from .validators import validate_youtube_url
 
 logger = logging.getLogger(__name__)
